+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - `src`: Fix access to private structures
 - `src/common`: Add missing option `-P` for hook calls
 - `src/libtman`: Replace struct with typedef
+- `src`: Move library limits into one place
 
 ### Deleted